split DTO.ts into separate files
