+++ polymorphia-frontend/app/(logged-in)/hall-of-fame/HallOfFameService.ts
@@ -1,6 +1,6 @@
 /* eslint-disable */
 
-import { HallOfFameResponseDTO } from "@/interfaces/api/DTO";
+import { HallOfFameResponseDTO } from "@/interfaces/api/hall-of-fame";
 
 export const studentNames = [
   "Gerard Małoduszny",
