+++ polymorphia-frontend/components/course/event-section/points-summary/types.ts
@@ -1,7 +1,5 @@
-import {
-  AssignedItemResponseDTO,
-  PointsSummaryDetailsResponseDTO,
-} from "@/interfaces/api/DTO";
+import { PointsSummaryDetailsResponseDTO } from "@/interfaces/api/course/points-summary";
+import { AssignedItemResponseDTO } from "@/interfaces/api/reward/assigned";
 import { ModalProps } from "../../../modal/types";
 import { RefObject } from "react";
