+++ polymorphia-frontend/components/equipment/types.ts
@@ -1,7 +1,7 @@
 import {
   EquipmentChestResponseDTO,
   EquipmentItemResponseDTO,
-} from "@/interfaces/api/DTO";
+} from "@/interfaces/api/equipment";
 
 export interface EquipmentSectionWrapperProps {
   items: EquipmentItemResponseDTO[];
