+++ polymorphia-frontend/components/hall-of-fame/desktop/HallOfFameCardDesktop.tsx
@@ -3,7 +3,7 @@ import HallOfFamePointsSummary from "@/components/hall-of-fame/general/HallOfFam
 import HallOfFameUserData from "@/components/hall-of-fame/general/HallOfFameUserData";
 import UserPoints from "@/components/user-points/UserPoints";
 import "./index.css";
-import { HallOfFameRecordDTO } from "@/interfaces/api/DTO";
+import { HallOfFameRecordDTO } from "@/interfaces/api/hall-of-fame";
 import { HallOfFameContext } from "@/components/providers/hall-of-fame/HallOfFameContext";
 import { useContext } from "react";
 import { filterXpDetails } from "@/components/providers/hall-of-fame/utils/filterXpDetails";
