+++ polymorphia-frontend/components/home/LoginForm.tsx
@@ -5,7 +5,7 @@ import { useForm } from "@tanstack/react-form";
 import { useMutation } from "@tanstack/react-query";
 import ButtonWithBorder from "@/components/button/ButtonWithBorder";
 import NavigationArrow from "@/components/slider/NavigationArrow";
-import { LoginDto } from "@/interfaces/api/DTO";
+import { LoginDto } from "@/interfaces/api/login";
 import { FieldInfo } from "@/components/form/FieldInfo";
 import { loginSchema } from "@/components/form/schema";
 import AuthService from "@/services/AuthService";
