+++ polymorphia-frontend/components/speed-dial/modals/GroupPickingModal.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { UserDetailsDTO } from "@/interfaces/api/DTO";
+import { UserDetailsDTO } from "@/interfaces/api/user";
 import Modal from "@/components/modal/Modal";
 import { useEffect, useState } from "react";
 import ButtonWithBorder from "@/components/button/ButtonWithBorder";
