+++ polymorphia-frontend/components/speed-dial/modals/types.ts
@@ -1,4 +1,4 @@
-import { EventType } from "@/interfaces/api/DTO";
+import { EventType } from "@/interfaces/api/course";
 
 export interface SpeedDialModalProps {
   gradableEventId: number | undefined;
