+++ polymorphia-frontend/interfaces/api/course/index.ts
@@ -1,0 +1,28 @@
+export type EventType = "ASSIGNMENT" | "PROJECT" | "TEST";
+
+export interface EventSectionResponseDTO {
+  id: number;
+  type: EventType;
+  name: string;
+  orderIndex: number;
+}
+
+export interface BaseGradableEventResponseDTO {
+  id: number;
+  type: EventType;
+  name: string;
+  topic?: string;
+  orderIndex: number;
+}
+
+export interface StudentGradableEventResponseDTO
+  extends BaseGradableEventResponseDTO {
+  gainedXp?: string;
+  hasReward: boolean;
+  isLocked: boolean;
+}
+
+export interface InstructorGradableEventResponseDTO
+  extends BaseGradableEventResponseDTO {
+  ungradedStudents: number;
+}
