+++ polymorphia-frontend/interfaces/api/course/points-summary/index.ts
@@ -1,0 +1,14 @@
+import { AssignedItemResponseDTO } from "../../reward/assigned";
+
+export interface PointsSummaryResponseDTO {
+  gained: PointsSummaryDetailsResponseDTO;
+  flatBonus: PointsSummaryDetailsResponseDTO;
+  percentageBonus: PointsSummaryDetailsResponseDTO;
+  total: PointsSummaryDetailsResponseDTO;
+}
+
+export interface PointsSummaryDetailsResponseDTO {
+  title: string;
+  gainedXp: string;
+  assignedItems?: AssignedItemResponseDTO[];
+}
