+++ polymorphia-frontend/interfaces/api/course/project/index.ts
@@ -1,0 +1,7 @@
+export interface ProjectVariantResponseDTO {
+  id: number;
+  name: string;
+  categoryName: string;
+  shortCode: string;
+  imageUrl: string;
+}
