+++ polymorphia-frontend/interfaces/api/equipment/index.ts
@@ -1,0 +1,15 @@
+import { ItemResponseDTO, ChestResponseDTO } from "../reward";
+import {
+  ItemAssignmentDetailsResponseDTO,
+  ChestAssignmentDetailsResponseDTO,
+} from "../reward/assigned";
+
+export interface EquipmentItemResponseDTO {
+  base: ItemResponseDTO;
+  details: ItemAssignmentDetailsResponseDTO[];
+}
+
+export interface EquipmentChestResponseDTO {
+  base: ChestResponseDTO;
+  details: ChestAssignmentDetailsResponseDTO;
+}
