+++ polymorphia-frontend/interfaces/api/hall-of-fame/index.ts
@@ -1,0 +1,14 @@
+import { UserDetailsDTO } from "../user";
+
+export interface HallOfFameResponseDTO {
+  content: HallOfFameRecordDTO[];
+  page: {
+    pageNumber: number;
+    totalPages: number;
+  };
+}
+
+export interface HallOfFameRecordDTO {
+  userDetails: UserDetailsDTO;
+  xpDetails: Record<string, string>;
+}
