+++ polymorphia-frontend/interfaces/api/login/index.ts
@@ -1,0 +1,4 @@
+export interface LoginDto {
+  email: string;
+  password: string;
+}
