+++ polymorphia-frontend/interfaces/api/markdown/index.ts
@@ -1,0 +1,3 @@
+export interface MarkdownResponseDTO {
+  markdown: string;
+}
