+++ polymorphia-frontend/interfaces/api/user/index.ts
@@ -1,0 +1,9 @@
+//TODO: handle which one is optional
+export interface UserDetailsDTO {
+  studentName: string;
+  animalName: string;
+  evolutionStage: string;
+  group: string;
+  imageUrl: string;
+  position: number;
+}
